Add the ability to recognize new binder event.
Extend the ability to binder_parser.html
with

 registerPrefixEventHandler('binder', ...
 registerEventHandler('binder_event_end',

The binder prefix matches all events not defined explicitly and the
binder_event_end closes whatever started previously by the binder
prefix. These provide an extensible support to the the binder events
and is able to recognize new event types not yet defined.